+++ lib/src/main/java/com/diffplug/spotless/java/ForbidWildcardImportsStep.java
@@ -18,17 +18,17 @@ package com.diffplug.spotless.java;
 import com.diffplug.spotless.FormatterStep;
 import com.diffplug.spotless.generic.ReplaceRegexStep;
 
-/** Removes any wildcard import statements. */
-public final class RemoveWildcardImportsStep {
+/** Forbids any wildcard import statements. */
+public final class ForbidWildcardImportsStep {
 
 	/**
 	 * Matches lines like 'import foo.*;' or 'import static foo.*;'.
 	 */
 	private static final String REGEX = "(?m)^import\\s+(?:static\\s+)?[^;\\n]*\\*;\\R?";
-	private static final String NAME = "removeWildcardImports";
+	private static final String NAME = "forbidWildcardImports";
 	private static final String ERROR = "Do not use wildcard imports (e.g. java.util.*) - replace with specific class imports (e.g. java.util.List) as 'spotlessApply' cannot auto-fix this";
 
-	private RemoveWildcardImportsStep() {}
+	private ForbidWildcardImportsStep() {}
 
 	public static FormatterStep create() {
 		return ReplaceRegexStep.lint(NAME, REGEX, ERROR);