+++ plugin-gradle/src/main/java/com/diffplug/gradle/spotless/JavaExtension.java
@@ -36,12 +36,12 @@ import com.diffplug.spotless.FormatterStep;
 import com.diffplug.spotless.extra.java.EclipseJdtFormatterStep;
 import com.diffplug.spotless.generic.LicenseHeaderStep;
 import com.diffplug.spotless.java.CleanthatJavaStep;
+import com.diffplug.spotless.java.ForbidWildcardImportsStep;
 import com.diffplug.spotless.java.FormatAnnotationsStep;
 import com.diffplug.spotless.java.GoogleJavaFormatStep;
 import com.diffplug.spotless.java.ImportOrderStep;
 import com.diffplug.spotless.java.PalantirJavaFormatStep;
 import com.diffplug.spotless.java.RemoveUnusedImportsStep;
-import com.diffplug.spotless.java.RemoveWildcardImportsStep;
 
 public class JavaExtension extends FormatExtension implements HasBuiltinDelimiterForLicense, JvmLang {
 	static final String NAME = "java";
@@ -152,8 +152,14 @@ public class JavaExtension extends FormatExtension implements HasBuiltinDelimite
 		addStep(RemoveUnusedImportsStep.create(formatter, provisioner()));
 	}
 
+	@Deprecated
 	public void removeWildcardImports() {
-		addStep(RemoveWildcardImportsStep.create());
+		System.err.println("`removeWildcardImports()` replaced by `forbidWildcardImports()`");
+		forbidWildcardImports();
+	}
+
+	public void forbidWildcardImports() {
+		addStep(ForbidWildcardImportsStep.create());
 	}
 
 	/** Uses the <a href="https://github.com/google/google-java-format">google-java-format</a> jar to format source code. */