+++ plugin-maven/src/main/java/com/diffplug/spotless/maven/java/ForbidWildcardImports.java
@@ -16,13 +16,13 @@
 package com.diffplug.spotless.maven.java;
 
 import com.diffplug.spotless.FormatterStep;
-import com.diffplug.spotless.java.RemoveWildcardImportsStep;
+import com.diffplug.spotless.java.ForbidWildcardImportsStep;
 import com.diffplug.spotless.maven.FormatterStepConfig;
 import com.diffplug.spotless.maven.FormatterStepFactory;
 
-public class RemoveWildcardImports implements FormatterStepFactory {
+public class ForbidWildcardImports implements FormatterStepFactory {
 	@Override
 	public FormatterStep newFormatterStep(FormatterStepConfig config) {
-		return RemoveWildcardImportsStep.create();
+		return ForbidWildcardImportsStep.create();
 	}
 }